requirements.txt
blinker==1.4
click==8.0.1
colorama==0.4.4
DateTime==4.3
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.1
Flask-Limiter==1.4
Flask-Mail==0.9.1
Flask-PyMongo==2.3.0
Flask-WTF==0.15.1
gunicorn==20.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
limits==1.5.1
MarkupSafe==2.0.1
pymongo==3.11.4
python-dotenv==0.18.0
pytz==2021.1
six==1.16.0
Werkzeug==2.0.1
WTForms==2.3.3
zope.interface==5.4.0
